requirements.txt
# Core dependencies
python-dotenv==1.0.0
supabase==2.3.4
gotrue>=1.3,<2.9  # Pin to avoid proxy parameter conflict with httpx 0.25.2

# Anthropic API
anthropic==0.39.0

# RSS feed parsing
feedparser==6.0.11

# HTML parsing and scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# HTTP requests
requests==2.31.0
httpx<0.26,>=0.24

# Vector embeddings (OpenAI for text-embedding-3-small)
openai>=1.12.0

# Data handling
pydantic==2.6.1

# Utilities
python-dateutil==2.8.2

# LinkedIn automation dependencies
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
flask==3.0.0
tenacity==8.2.3
